backend_analysis/news_crawler/requirements.txt
requests
fastapi
uvicorn
python-dotenv
apscheduler
